WP.com block editor: Ignore token on cross-site cookies (#14565)
Our filters hooking into set_auth_cookie and set_logged_in_cookie were expecting 6 params. The 6th param $token was added in WP 4.9.0, so sites using older versions of WP are running into errors when using Jetpack 8.1.1:

This update removes that param from the hook callback since it was not used, which effectively fixes the issue with sites using older versions of WP.

It also changes the filter to determine if logged in cookies are secure from secure_auth_cookie to secure_logged_in_cookie which is the filter used in the original pluggable function. This might solve issues reported in #14385, but we've not found yet exact steps to reproduce them, so we could need more changes.