Updated the rendering of the styles to use the `style` attribute
When rendering the block colours on the server, we were using an
`inline_style` that rendered as a style tag. This meant that each style
tag would override the last, and we were also using the wrong selector,
so the colours weren't being applied on the front end.

This change adds the styles to each button as an inline style attribute,
so they're unique to each instance.