OpenTable: Allow the various block styles to be centre aligned (#14633)
* Allow the various block styles to be centre aligned

The centre alignment wasn't working as the iframe is set to 100% and its
content was left-aligned. The width of the iframe is getting set by the
OpenTable embed JS code, so in order to override it, we need to use
`!important`. That's not ideal, but I think as it's isolated, it should
be ok.

* Make sure the alignment styles work in the editor

Co-authored-by: Ben Dwyer <ben@scruffian.com>