core testcase: share more code with the isolated base testcase
Summary:
There is not reason the core base testcase should not call the parent tearDown method, like all other testcases do.
Core's cache flush is simplified, and the redundant calls compared to the base isolation testcase are removed.

This makes the environment slightly incompatible with the alloptions tests, and reveals a test that should be failing, but reduces the testcase maintenance burden. Fixes are provided for the tests

Test Plan: TC

Tags: #touches_jetpack_files

Differential Revision: D38799-code

This commit syncs r202837-wpcom.