Fix error when pasting embed codes
We were accidentally relying on an error being thrown during the
onChange handler, when an embed code was pasted into the input. The
error meant that the event bubbled up and the onSubmit event handled the
input correctly.

This change sets the saveTransform function to check if the current
input looks like an embed. If it does, then it prevents it from being
treated like a token, and the onChange handler is never called.