Ensure default block classname is used as the root of other classes
We've been using the `className` prop to compose other class names for
the layout of the block edit component. e.g. `${ className }-overlay`.

As pointed out by @jeherve in #14623, this breaks if additional CSS
classes are added to the block in the advanced settings.

This change, uses `getBlockDefaultClassName` to get the default class
name to use instead.