Ensure className is preserved
As the HOC was used as part of a filter, the className was getting set
by the last filter that was run. This was a lucky accident for OpenTable
as it meant that it wasn't affected by the bug being fixed in #14732

This change passes through the className prop when the block names don't
match, which means that if it's been set by a previous filter, it is
preserved.