SubmitButton: Make it use the theme palette (#14599)
The `SubmitButton` component was setting the `customButtonTextColor`
(or the `customButtonBackgroundColor`) attribute whenever a colour was
set. This would override any colour classes set to use the theme
palette, like `has-primary-color`. As a result the colours wouldn't
update as the theme was changed.

This also meant that a few components were only rendering the button
using the custom colour attributes, and so would suffer from the same
problem.

Additionally, the Calendly attribute validation was not allowing colour
names like `primary`. For now I've removed the validation as everything
is properly escaped, but in future we should probably create a better
regex to check for hex values or hyphenated lowercase words.

refactor recurring payments class names to make the code more readable

use links for the recurring payments button, so it matches the other uses of this block

Add positional markers to the arguments to sprintf

Add check for submitButtonClasses being set

Remove className attribute from SubmitButton

The classes being passed into SubmitButton appear to be redundant. Thes
removes them and explicitly passes just the props needed.

Make attribute check clearer

Format the attribute definition consistently

Explicitly pass the required props to SubmitButton