Calendly: Changed the embed, to remove the inline styles (#14579)
* Changed the embed, to remove the inline styles

The inline styles added to the parent container by the Calendly script
were interferring with the CSS used for aligning elements to the left
and right on the desktop view of Twenty Twenty (and so probably other
themes)

An undocumented option to the init function stops these styles being
added, and so this change swaps the embed code to set that option, as
well as tweak the CSS to make a best effort at laying out the embed when
aligned to the sides.

* Scoped the height setting to the inline style

* Prevent the server side rendering in the admin

The render_callback was being called to get the content of the page in
the admin, and this was causing the block embed to render full screen
when it was inline mode. This short circuits the render callback in the
admin.

* Changed the embed, to remove the inline styles

The inline styles added to the parent container by the Calendly script
were interferring with the CSS used for aligning elements to the left
and right on the desktop view of Twenty Twenty (and so probably other
themes)

An undocumented option to the init function stops these styles being
added, and so this change swaps the embed code to set that option, as
well as tweak the CSS to make a best effort at laying out the embed when
aligned to the sides.

* Simplify the block ID code, and remove is_admin check

We've discovered using `wp_unique_id`, so it seemed sensible to use that
here.

There was a check for is_admin in the render callback, because of an
error with the post being rendered as the editor loaded. I can't
replicate this now, so I've removed the check.

* Scoped the height setting to the inline style

* Prevent the server side rendering in the admin

The render_callback was being called to get the content of the page in
the admin, and this was causing the block embed to render full screen
when it was inline mode. This short circuits the render callback in the
admin.

* fix rebase

* properly escape the block_id

Co-authored-by: Ben Dwyer <ben@scruffian.com>