[not verified] Fix the overlay when the block is left or right aligned (#14820)
In WordPress/Gutenberg#19593 a wrapper `div` was removed, which had a
`position` value of `relative`. Our overlay `div` to prevent the iframe
from being interacted with, relies on a parent element being
`positino: relative`. This is still the case when the block is in the
main flow/centre aligned, but when the block is left or right aligned,
it can't be selected, and clicks are handled by the iframe.