Prevent noticons from being enqueued in AMP mode
Summary:
This patch disables enqueuing of assets for some non-functional features
when in AMP mode.

Noticons is a very large CSS asset, and so it prevents a lot of other CSS from
being enqueued. A review of the places we use it showed that it was not actually
being used by any functionality that works in AMP mode.

Affected features:
- Comment Likes (non-functional in AMP)
- Geo Location (not used on most pages - might be useful to port `wpcom_vip_load_geolocation_styles_only_when_needed()` to all sites)
- Notifications UI (non-functional in AMP)
- admin-bar - (non-functional in AMP)

Test Plan:
Load a page in Transitional mode

Ensure that noticons-css is not enqueued, and that little or no CSS is blocked by AMP due to being over the 50kb max

[x] - ensure that features whose CSS is not being enqueued are not actually used

Reviewers: josephscott, #devops_team, davidbinovec, goldsounds

Subscribers: davidbinovec

Tags: #touches_jetpack_files

Differential Revision: D38340-code

This commit syncs r203534-wpcom.