Release Build: ensure the first built release branch is cleaned
We don't want that built branch to include things like /packages, because those then remain behind later on when you update the built branch.

By purging it from most things the first time around, we ensure that when you update it later on and you rsync the local built version (which is stripped of files from .svnignore) to your remote built version, you do not bring files listed on .svnignore back.