Static binding fix for `Jetpack` class.
`Jetpack` singleton class could not be properly extended due to lack of late static binding.
Test class `WP_Test_Jetpack` relies on `MockJetpack` which extends `Jetpack`, but under the hood it still works with the parent `Jetpack` instance instead of an `MockJetpack` object.
This means that the test does not create its own environment to modify, but modifies the global environment and affects other tests that run after it.

The commit provides three changes:
1. The `MockJetpack` class gets its own `$instance` property that holds a `MockJetpack` object.
2. `MockJetpack` also redefines the `init()` method to create a `MockJetpack` instance instead of a `Jetpack` one.
3. Class `Jetpack` is modified to let `connection()` method to work with a correct instance of the calling class.