Infinite Scroll: Remove page caching; append new content as DOM elements
Currently the Infinite scroll receives HTML string and updates the page
using wrapper.innerHTML = {old-content} + {new-content}.

This creates multiple issues, the most glaring problem being that any and all
event listeners originally bound on any of the elements within the wrapper
element are destroyed whenever a new IS page is rendered.

The current system also tries to improve the performance of large ISs by
removing all DOM nodes from IS pages that are outside of the current viewport.
When this happens the existing content is cached - again as a HTML string.
Unsurprisingly, that leads an issue identical to the above.