Customizable page template
Summary:
This patch adds two new endpoints: One that receives the name of a page template, the locale and a vertical name and returns a json with metadata, including the code of the template after applying the content pack that matches the received parameters.
The other endpoint receives the same kind of parameters, plus a theme, and returns the html of how the customized template would look like if used in combination with a theme (intended to be used as a preview of a site before changing to a new page template)

Phabricator delenda est.

Test Plan:
Right now there's only one template that can be used as an example: the home page template include with Rockfield (the theme).

So you can check that both API endpoints work fine and customize the template using one of the two vertical content packs that have been prepared as an example ("football" and "gym").

https://[private link]

https://[private link]

Change the vertical in both endpdoints and check the content actually changes

Reviewers: jonsurrell, simison, ramonopoly, noahtallen, vindl

Reviewed By: jonsurrell

Tags: #touches_jetpack_files

Differential Revision: D40475-code

This commit syncs r204594-wpcom.