Restore the `ab_connect_banner_green_bar` option to the valid options list to allow safe removal when upgrading to Jetpack 8.4.
In PR #14969 the `ab_connect_banner_green_bar` variable was removed from the list of valid option names. This clashes with the upgrade code which checks for the option's existence when Jetpack is upgraded in order to remove it.

Since the option is not on the list of valid options, a PHP warning will be triggered.

This PR restores the variable in the valid options list to make sure this won't throw any errors.