WP for Teams: add is_wpforteams_site option to sites endpoint (#15233)
* WP for Teams: add is_wpforteams_site option to sites endpoint

Summary: In this patch, we add an option `is_wpforteams_site` to all `sites/` REST API endpoints. We need that so we can distinguish between standard WP.com sites and WP for Teams sites in Calypso and other places.

Test Plan: Apply the patch. Open Calypso and inspect the `sites/` endpoint response. It should contain the `options` field with `is_wpforteams_site` option either `true` or `false` depending on whether the site was created with the `start/wp-for-teams` signup flow or not.

Reviewers: #lighthouse_team, migueluy

Reviewed By: #lighthouse_team, migueluy

Subscribers: migueluy

Tags: #touches_jetpack_files

Differential Revision: D41160-code

This commit syncs r205198-wpcom.

* Add is_wpforteams_site method

* Add is_wpforteams_site to site_options_format

Co-authored-by: Rastislav Lamo <lamosty@automattic.com>