Fix follow button alignment in /blog
Summary:
Issue reported in p1585043548379400-slack-CFFF01Q4V

The Follow button's alignment is off. This patch fixes it.

**LOGGED IN**

**Before**

{F5553176}

**After**

{F5553177}

Mobile view:

{F5660046}

**LOGGED OUT**

**Before**

{F5659152}

**After**

{F5659153}

Mobile view:

{F5660047}

Another small improvement is to increase the left padding for the placeholder text  for the email input box. Currently, the padding is low and the placeholder text and email input are almost brushing against the walls of the input box. A little bit of padding helps make it look better:

**Before**

{F5659408}

**After**

{F5659409}

Test Plan:
1. Sandbox wordpress.com
2. Go to https://wordpress.com/blog and open any article,
3. Scroll down to the Email newsletter section and verify that the follow button's alignment looks okay. Try while logged in and logged out to wpcom
4. Check the subscription widget on the other locale blogs that we have- ja.[private link] & fr.[private link]. Remember to sandbox the domains. On each of these blogs, open any article, scroll down and verify that the subscribe widget's follow button alignment is right while logged in and logged out.

Reviewers: #martech_team, jaswrks

Reviewed By: #martech_team, jaswrks

Subscribers: louislaugesen, jaswrks

Tags: #touches_jetpack_files

Differential Revision: D40953-code

This commit syncs r205322-wpcom.