Skip the contact form spam recheck if there's no information saved to use during the recheck.
When you click "Check for Spam" on the Jetpack Contact Form page, it pulls the _feedback_akismet_values post meta for each feedback and uses that to check whether the submission is spam.

However, the _feedback_akismet_values is deleted after a feedback is 15 days old by the daily_akismet_meta_cleanup() function. This is reasonable, and it is also reasonable not to check feedback older than 15 days old for spam, since spam checking relies on real-time data.

This pull request skips the spam recheck if there is no daily_akismet_meta_cleanup post meta.