Store Product List: Stop setting attributes on the Store_Product class object to fill in data for the API responses
Summary:
This makes updates to the methods in Store_Product_List that were previously setting different attributes on the Store_Product to generate responses for the API endpoints.
Instead, this is now generating generic objects to be used as DTO's for the purpose of the API.

This did require updating some methods outside of the Store_Product_List to make sure they were using the methods that returned Store_Products[] vs the DTO's where necessary.

Test Plan:
Built in tests.

I also applied D41723, which can be used to generate JSON output from the API endpoints that use the Store_Product_List and can be compared before and after applying this patch.

Reviewers: #florin_team, #payments_team, #marketing_team, #jetpack_crew_team

Tags: #touches_jetpack_files

Differential Revision: D41787-code

This commit syncs r206104-wpcom.