Recipe shortcode: on AMP, change the Print link (#15492)
* Recipe shortcode: on AMP, change the Print link

Simply print the page,
instead of only the recipe.
It'd be nice if it was possible to only print
the recipe.
But that's done with a jQuery plugin.
It doesn't look possible in an <amp-script>,
as the jQuery plugin creates an <iframe>,
which <amp-script> doesn't allow.

* Change esc_html_x() to esc_html__(), as context doesn't seem needed

The context of 'recipe' was helpful for
printing only the recipe.
But the context doesn't seem needed for 'Print Page'.

* Attempt to fix failed unit tests

Instead of creating an actual post,
create an object with post content.

* Attmempt to fix Code Climate issue

Function kses_tags has a Cognitive Complexity of 6 (exceeds 5 allowed). Consider refactoring

* Remove duplicate dequeuing

I intended to make this wp_dequeue_style(),
but it looks like that might not be needed.