Comments: do not create comments with the comment type just yet
This is a follow-up on #15494.

Let's not create "comment" type comments that are then displayed in areas handled by themes just yet. A lot of themes are not yet ready to handle comments that have a comment type other than an empty string.
We can apply this change once WordPress 5.5. ships, since there will then be a dev note published helping theme authors update their theme to take this into account.

Until then, let's only support the new comment type when pulling comments, and let's use the comment type in comments that are displayed via Jetpack, such as in the Carousel.