Onboarding: saving the site creation flow as a site option
Summary:
Heyho,

This patch saves the value of the `site_creation_flow` option as a site option.

We're doing this so we know which onboarding flow the user completed to create a particular site.

More specifically, we'd like to identify sites created in the Gutenboarding signup flow so we can customize the editor UI.

41300-gh-wp-calypso#issuecomment-617065342

Test Plan:
Run https://github.com/Automattic/wp-calypso/pull/41349

Create a site via `/new`.

When you finally land in the editor, filter the XHR network requests in your browser's console by `me/sites`

Check that the option is set.

Reviewers: #ganon_team, #luna_team, philipmjackson

Reviewed By: #ganon_team, philipmjackson

Subscribers: philipmjackson

Tags: #touches_jetpack_files

Differential Revision: D42191-code

This commit syncs r206374-wpcom.