Instagram Widget: Fix the checks for a valid Keyring token https://github.com/Automattic/wpcomsh/pull/452
* Instagram Widget: Fix the checks for a valid Keyring token

This is to be deployed in conjunction with D42120-code. The Keyring
token can be deleted from other places, so these changes are to reset
the token ID in the widget if the one stored is no longer valid.

Testing is the same as for D42120-code. A connection to Instagram should
be made via the widget, and disconnected via Calypso. The widget should
then be reset rather than displaying an error about the API.

* use a different route to check the token

Co-authored-by: Ben Dwyer <ben@scruffian.com>