Moved the logic to gate WPCOM blocks to a shared file (#14426)
* [not verified] [not verified] # This is a combination of 5 commits.

[not verified] Moved the logic to gate WPCOM blocks to a shared file

The logic to gate the blocks to paid plans on WPCOM is repeated for the
OpenTable and Calendly blocks. It is likely to be needed in future too,
so this change moves the repeated to logic to the `Jetpack_Gutenberg`
library.

Reworked incorporating ideas from #14430 and #14431

Changed availability logic so a block is always registered

Based on the suggestion by @simison, this changes the logic around block
availability, so that a block can always be registered, but its
availability can be set independently and used to prevent dynamic blocks
from rendering.

Adjusted the call to `remove_extension_prefix` because of load order

The compatibility function `wp_startswith` isn't loaded until the
`plugins_loaded` hook, but plugins sometimes call
`jetpack_register_block` before that. This caused an exception in the
`remove_extension_prefix` call.

This moves that call to the point just before the value is used,
ensuring that the function is available.

Refactored to work with the existing `Jetpack_Plan::supports` method

This is missing the WPCOM part at the moment.

* [not verified] Added the first version of the plan check for WPCOM

* [not verified] Stop the availability check for non-jetpcak blocks

A core test was failing for core/video because Videopress registers a
server render_callback. This is the exception and we don't want to do
avaavailability checks on non-jetpack blocks, so this change adds a
check on the slug as the block is registered.

* [not verified] Corrected the comments on the availability methods

* [not verified] Updated logic to supply minimum plan slug for the WPCOM nudge

* [not verified] Simplified the code to wrap the `render_callback`

* Rebased to include the frontend upgrade nudge

This also updates the logic so the availability check is done in
`jetpack_register_block`

* Change call to wp_startswith to allow the function to be called in init

The wp_startswith function isn't available on wpcom until
plugins_loaded. This changes the calls to it, to the strpos equivalent.

* Move the availability check so it doesn't happen in the init hook

On wpcom, the availability check was getting the plans form the databse
and that was causing havoc with the isolated tests. I think this should
fix it.

* Simplify the plan_check logic

Co-authored-by: Jeremy Herve <jeremy@jeremy.hu>

* Simplify the 'if' statement

Co-authored-by: Jeremy Herve <jeremy@jeremy.hu>

* Add some documentation

* Added an extra line to the docs about where to find plan data.

* Documentation updates

Co-authored-by: Jeremy Herve <jeremy@jeremy.hu>
Co-authored-by: Ben Dwyer <ben@scruffian.com>