Allow `spotify:` text to be embedded when not preceded by HTML tag.
Because of how the content is split with `wp_html_split()`, the `spotify:` embed text would sometimes not be the first 8 characters of an `$element` even if it were on its own line. The regex in the `preg_replace_callback()` on L83 will only match if it's at the beginning of a line so checking the first 8 characters of `$element` is redundant and caused some valid embed texts to be skipped.

This change simply checks to see if the `$element` contains the `spotify:` string before attempting the `preg_replace_callback()`.