Connection: increase timeout on the token request (#15764)
* Connection: increase timeout on the token remote request

We're seeing timeouts on the token reqquest in the logs. Let's try increasing the timeout
from the default 5s to 10s.

* Connection: use the filter to increase the token request timeout

Use the 'http_request_timeout' filter to increase the token request
timeout to 30 seconds.

* Connection: add priority to http_request_timeout hooks

Add a priority value of PHP_INT_MAX-1 to the add_filter('http_request_timeout' ...)
and remove_filter('http_request_timeout' ...) calls. This will help prevent other
plugins from accidentally changing the timeout value that we need for the token request.