Editor: Update switch to classic editor link (#15644)
Update Switch To Classic Editor link

When the WP.com classic editor is deprecated the Switch to Classic Editor link needs to go to the WP admin classic editor. This includes activating the Classic Editor plugin if it is available and resetting user options to ensure that the user actually ends up being given the WP admin classic editor. The "Switch To Classic Editor" link will only be shown when:

- Loading an Atomic or Jetpack site in WPcom Calypso and using the iFramed block editor
- The editor/after-deprecation query string param is present (Temporary feature flag).
- The classic editor plugin is installed but not active.
- User has permission to activate plugins e.g. admins.

Co-authored-by: Glen Davies <glen.davies@a8c.com>
Co-authored-by: Ben Dwyer <ben@scruffian.com>