Update visibility of Switch To Classic Editor link
As we are rolling out the editor deprecation in to users in stages, we need to allow for the original conditions for the display of the Switch To Classic Editor link.

If the user is not part of the editor deprecation rollout group the only condition required is that they are in the iframed block editor.

If they are in the rollout group then we need to also check for the presence of the classic editor plugin, that its inactive and the user has permissions to activate plugins.

Due to difficulty testing iframed block editor locally ability to force display with query string param has been added as well.