REST File upload: Add custom error codes
Summary: This patch addresses the issue that in non-English we are not showing differenciated error messages when the upload cannot happen because of the file size (too big, over quota etc.) The client-side fix for this is https://github.com/Automattic/wp-calypso/pull/42921

Test Plan: Test along with https://github.com/Automattic/wp-calypso/pull/42921, for a full media library there should be the right error message when the upload fails because of lack of space.

Reviewers: yslavchev, zinigor

Reviewed By: yslavchev

Subscribers: yslavchev

Tags: #touches_jetpack_files

Differential Revision: D44299-code

This commit syncs r208397-wpcom.