Autoloader: handle an updating plugin
The Autoloader generates the in-memory classmap when the first autoloader package
is loaded. When a plugin is updated, the path to a package class could have changed,
making the in-memory classmap point to an invalid file location.

Reset the autoloader when a plugin with the v2.x autoloader is updated so that the classmap
is kept up to date.