Add checks for objects and remove unneeded call
Due to the way React useLayoutEffect works, the scanPage function is
called multiple times, with the component in various states of ready.

This ends up causing problems if one of the states is not rendered how
it expects, this change adds a couple of checks to make sure it is in a
valid state or return.

Additionally, the initial scanPage call at the end is no longer needed
and was removed.