Update block button url with embed params
After switch to using the block button and passing through the url attribute, the embed params for background, text and primary color were lost. The original approach to parsing embed code strips the url back to basics and only adds the query string params when generating markup.

This change simply replaces the base url with the one containing query string params as this had the least impact on existing approach.