Calendly Block: Replace SubmitButton with new button block (#15863)
Replace Calendly SubmitButton with shared button

The new shared button offers a few extra features and an opportunity to make
multiple components more consistent. As such, the SubmitButton in the Calendly
block is being replaced. 

This fixes a minor bug where the Calendly spinner for an inline style widget isn't hidden when there are multiple Calendly blocks on the page.

Switching to the shared Button block also fixes an issue where custom text color wasn't shown in editor.

Co-authored-by: Copons <Copons@users.noreply.github.com>
Co-authored-by: Jeremy Herve <jeremy@jeremy.hu>