XMLRPC API: Introducing `jetpack_connection` flag.
The new flag `jetpack_connection` indicates that the site is connected to WordPress.com.
The existing flag `jetpack` assumes that Jetpack plugin itself is installed.

Since the endpoint itself is located in Jetpack anyway, both flags are hardcoded as `true`.
At the moment they are only needed to make the endpoint compatible with WP.com API.