Autoloader: fix the fatal that occurs during plugin update
Fix the fatal that occurs during plugin update. The fatal occurs because a function in
the updated version of autoload_functions.php was called. However, that file had already been
loaded, so the updated version of  the file isn't available.

Instead, put the upgrader_post_install callback function in the autoload_functions.php file
and call set_up_autoloader in the existing file to reset the autoloader after a plugin update.