Memberships: Add a route for generating products. (#16242)
Part of adding support for the Donations block will be automatically generating default plan products for the three tabbed options: one-time, monthly, and yearly. This adds a route to the memberships endpoint for POSTing that a certain type of product have its defaults generated (and returned). The only type to be immediately supported will be the donation type.