Autoloader: use paths to identify plugins instead of the directories
Currently, the autoloader identifies each plugin using its directory name.
The paths to the plugins' files are then generated using WP_PLUGIN_DIR. However, some
plugins, such as mu plugins, are not located in WP_PLUGIN_DIR. In those cases,
the autoloader fails to find the plugins' files.

To fix this, store the full path to each plugin's directory, and use that path
when obtaining the plugin's files.

Many plugins are initially identified with a plugin slug, such as the plugins obtained
with get_option['active_plugins']. For those plugins, we assume that they're in
WP_PLUGIN_DIR and add that to the plugins' paths.