Tiled Gallery Block: only available when Jetpack is connected
Fixes #16426

The Tiled Gallery block relies on Jetpack's Image CDN to resize the images on the fly, and consequently cannot be used on a site that's not publicly accessible, where the CDN cannot fetch the images.

With this change, the block will not appear when your site is not connected to WordPress.com.