Add migration to prevent language switch invalidation of block
When a block is created in English and the site switched to a second langauge before the block is migrated, the block will be deemed invalid.

This change will attempt to fall back to English defaults for the placeholder text in the saved shortcode that causes the invalidation. This will not change migrated attributes only the generated content from the deprecated block save function. It also will only cover blocks first created in English and using default text. If the block was created in a second language, there is no means to determine the language and therefore the default text for that translation to fall back to.