Connection Errors: Properly handle invalid JSON response.
If the endpoint `jetpack/v4/site` returns invalid JSON, the exception is thrown, but properties `message` and `code` of the error object don't exist.
It leads to an error displayed in the browser console, and the "Reconnect" error notice doesn't show up.
This commit fixes the issue, providing default error message and code, therefore the error notice will still show up in the dashboard, and suggestion reconnecting.