Admin Page: create new create interpolate element
This element acts as a wrapper for Core's own createInterpolateElement.

createInterpolateElement is available in WordPress 5.5 and in latest versions of the Gutenberg plugin, but it is not available in WordPress 5.4, which we still support.
This new component can consequently be used for now, until WordPress 5.5 becomes the minimum required version for Jetpack. When that happens we can deprecate this component and rely on createInterpolateElement directly.