Temporarily remove button block style when multiple RIDs set
OpenTable block’s button style isn’t compatible with multiple restaurant selections. Previously the button style was just blocked/switched to standard style. This was a bad UX and needed improving. Now when additional IDs are added, the style is removed and re-registered when switching blocks or removing all but one ID.