Sync: Fix nonce action string in theme edit sync
There is currently a mistake in the edit theme jetpack sync functionality that is preventing theme edited events from reaching the Activity Log.

The nonce field is being created like this in the theme editor:
wp_nonce_field( 'edit-theme_' . $stylesheet . '_' . $relative_file, 'nonce'  );

To test, Apply this patch, edit a theme and check that there is a theme modified event in the Activity Log.