Autoloader: change the functions.php file name and don't reset the version after update
During a plugin update, an error can be caused by opcache holding the old version of the
autoload_packages.php file when find_latest_autoloader tries to load the new version.
To prevent this, don't try to find the latest autoloader version; just continue using
the current autoloader.

Also, change the filename of the the autoload_functions.php file to autoloader_functions.php.