SSO: Getting rid of the user token.
The SSO login process performs an API request to WP.com (`jetpack.sso.validateResult` endpoint) to validate the authentication the results and get user information.

Jetpack tries to use `user_token` for that request, but at that point the user is not authenticated in WP, so `user_id` equals `0`, thus `blog_token` is used for authorization anyway.

That makes the commit janitorial with no functional changes.