Customizer: Tweak cookie widget edit button position (#16890)
There has been a little confusion over why the EU cookie banner widget
was shown within the Customizer. It is there as the user can configure a
number of the widget's properties and they'll need to be able to see the
results of those customizations.

This change just repositions the edit button that is added to the widget
within the Customizer to make it more obvious it is not their by mistake
and is editable.