Autoloader: try to detect when both v1.x and v2.x files are present
The autoloader can end up with some v1.x files and some v2.x files in this situation:
1) At least two plugins on a site have v2.x autoloaders.
2) A backup is restored, and the backup contains a version of a plugin with a v1.x autoloader. Therefore, the
   backup downgrades the autoloader for that plugin from v2.x to v1.x.
3) The autoloader files that are exclusive to v2.x remain in the plugin's directory. The file that v1.x and 2.x have
   in common, autoload_packages.php, is replaced with the v1.x version.

When this occurs, v2.x autoloaders on the site think that this plugin has a v2.x autoloader, but it actually runs a v1.x autoloader.

To fix this situation, detect that plugin has both v1.x and v2.x files and figure out which will run. Also, try to delete the unneeded files.