Tracking: Adds connection status check.
This is required now because the connection status check has been removed from the ToS package in #16967. This is a minimal change, but we should also take a look at how we are passing the Connection Manager object to the constructor of this class and remove the optional direct instantiation.