Docker: Make adding SFTP public key auth easy
The docker image being used for providing SFTP access, atmoz/sftp,
supports key-based authentication out of the box if you map public key
files to a particular directory in the container.

Set up a host directory mapping for that by default so all someone has
to do to make use of it is drop the public key file in the right place.